+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file. Dates are displayed in UTC +8.
 
 
+### v1.0.1
+> August 26, 2023
+- Added entrance animation for header & hero section.
+- Added animation for navigation menu.
+- Added message sent on contact form when success.
+
 ### v1.0.0
 > August 25, 2023
 - Added header with responsive navigator section on it.